+++ UniversityCMS/src/main/java/ua/foxminded/universitycms/repository/TeacherRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import ua.foxminded.universitycms.model.Teacher;
 
 @Repository
-public interface TeacherRepository extends JpaRepository<Teacher, Integer>{
+public interface TeacherRepository extends JpaRepository<Teacher, Long>{
 
 	@Query("SELECT COUNT(t) = 0 FROM Teacher t")
 	boolean checkIfEmptyTable();
